# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Deenly Backend API
  version: 1.0.0
  description: Stable v1 contract for frontend integration.
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        requestId:
          type: string
paths:
  /auth/register:
    post:
      summary: Register user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username, password, displayName]
              properties:
                email: { type: string, format: email }
                username: { type: string, example: abdullah_01 }
                password: { type: string, minLength: 8 }
                displayName: { type: string }
      responses:
        "201":
          description: Registered
  /auth/login:
    post:
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Logged in
  /auth/session/me:
    get:
      summary: Get authenticated session user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
  /users:
    get:
      summary: List users (stable profile surface)
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: User list
  /users/me:
    get:
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
    put:
      summary: Update current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated profile
  /users/{userId}:
    get:
      summary: Get user profile by id
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: User profile
  /profiles:
    get:
      deprecated: true
      summary: Deprecated alias for users list
      responses:
        "200":
          description: Profile list
  /posts:
    post:
      summary: Create post
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Post created
    get:
      summary: List posts
      parameters:
        - in: query
          name: postType
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50 }
        - in: query
          name: authorId
          schema: { type: integer }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: Post list
  /posts/{postId}:
    get:
      summary: Get post by id with metrics
      parameters:
        - in: path
          name: postId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Post details
  /feed:
    get:
      summary: Get ranked feed items (cursor pagination)
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50 }
        - in: query
          name: cursor
          schema: { type: string }
        - in: query
          name: postType
          schema: { type: string }
        - in: query
          name: followingOnly
          schema: { type: boolean }
        - in: query
          name: authorId
          schema: { type: integer }
        - in: query
          name: minCreatedAt
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Feed list
  /interactions:
    post:
      summary: Add interaction (benefited, reflect_later, comment)
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Interaction stored
  /interactions/view:
    post:
      summary: Track post view metrics
      security:
        - bearerAuth: []
      responses:
        "201":
          description: View event stored
  /interactions/post/{postId}:
    get:
      summary: Get interaction totals for a post
      parameters:
        - in: path
          name: postId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Interaction totals
  /reports:
    post:
      summary: Submit moderation report
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Report created
  /media/upload-signature:
    post:
      summary: Generate signed media upload URL
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Signed upload data
  /media/posts/{postId}/attach:
    post:
      summary: Attach raw uploaded media to post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: postId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Media attached
  /notifications:
    get:
      summary: List in-app notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications list
  /users/me/interests:
    get:
      summary: List current user interests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Interests list
    put:
      summary: Replace current user interests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Interests updated
  /users/me/sessions:
    get:
      summary: List refresh sessions for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sessions list
  /admin/tables/{table}:
    get:
      summary: Admin table viewer endpoint
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: table
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Table rows